# Hooks: Micro-optimize runHooks