--- a/package.json
+++ b/package.json
@@ -6,6 +6,7 @@
 	},
 	"devDependencies": {
 		"babel-core": "^6.26.0",
+		"benchmark": "^2.1.4",
 		"chalk": "^2.0.1",
 		"check-node-version": "^3.1.1",
 		"codecov": "^2.2.0",