--- a/packages/hooks/benchmark/index.js
+++ b/packages/hooks/benchmark/index.js
@@ -1,0 +1,18 @@
+const Benchmark = require( 'benchmark' );
+const hooks = require( '../' );
+
+const suite = new Benchmark.Suite;
+
+function myCallback() {}
+
+hooks.addFilter( 'handled', 'myCallback', myCallback );
+
+suite
+	.add( 'handled', () => {
+		hooks.applyFilters( 'handled' );
+	} )
+	.add( 'unhandled', () => {
+		hooks.applyFilters( 'unhandled' );
+	} )
+	.on( 'cycle', ( event ) => console.log( event.target.toString() ) )
+	.run( { async: true } );